Added a new syntax for verbose embeddings. This adds parentheses:
    $from(action_name);

Verbose embeddings without parentheses can make code difficult to read
because they force a space in the middle of an action embedding. There is a
tendency to associtate spaces with concatenation. Without syntax
highlighting to make it clear that the embedding type is a keyword, the
problem is especially bad. The danger is that a verbose embedding could be
read as an embedding of the embedding type:

    main := 'foo' $from some_action '\n';

With parentheses this statment reads much more clearly.

    main := 'foo' $from(some_action) '\n';