The compile functions should be const. We need to call them more than once to compile WV and WC code so they should not be modifying the AST.
Create the object field for the iterator at parse time with all the other field
creations. We need this once we start doing mutiple compiles. But still create
the type reference for the iterator variable at compile time since it is
dependent on the iterator call lookup and all call lookups are done at compile
time.

We also need to turn off the unscoping of the iterator variable so we can
support mutiple compiles. Eventually we will need a proper scoping solution and
this was just a little hack for one case.