Moved the compile-time InputStream classes (Pattern & Replacement parsing) to ctinput.cpp.
Split the overloaded functions pushBack to pushBackData and pushBackBuf. Made
all base functions virtual and moved the assert(false) versions into the super
classes.

Fixed a bug in FsmRun::sendBack. If the item going back is a named element then
we need to send back the uncomsumed part of buffer. Both p and pe were set to
the start of the current buffer. Just need to set pe = p. Also fixed the
comment.