The keyword 'cons' is now a synonym for 'construct'.
The keyword 'parser' is now a synonym for 'accum'. I was hesitant of this at
first due to it being only one letter away from 'parse'.

Now treating parser context as an object that needs to b passed to parse
statements of types that have context and passed to an accumulator (parser
object).