Install the runtime headers.
Improved the check that decides if we are in the source tree VS an installed
location. Previously, if there was a / in the path it was assumed to be source.
This doesn't work for cases where we specify the full path to the installed
location. Now we look for a slash, then for main.cc next to the binary.

refs #297.