Testing out a shadow tree.
The existing parse tree retains current functionality. It contains the shadow,
which is the data tree and will be exported by the parse. Testing method: when
the tree is submitted for parsing a copy into the shadow is done. When shifting
and reducing we need to manage the shadow tree's pointers. Basics works. No
ignore tokens.