Now translating only the parseTree->id to termDup from nonterminals. The data tree id always remains as the nonterminal.
The nested comm output needs to be adjusted because it is now functioning
correctly. A nonterm was sent as a comment and was not translated back to the
nonterm from the dup when the comment was attached. Now that translation
doesn't happen at all and so it is correctly a nonterminal.