a follow-ignore that appears to work
Duplicated the tokenRegion array with the preRegion taken from the same state.
This turns the region into a pair of values. The pre-region is tried first.
Eliminates the need to carry the pre-region anywhere in runtime state. Just
checked when setting cs. If present cs is set using the pre-region and ncs is
used for the real region. Not optimal space usage but that's okay for now.

Many test case failures. Need to verify thew new output is correct.