first checkin on no-kid-flags and no-ignore-dupes branch
Trying out an elimination of the kid flags and the duplicate ignore tokens. The
kid flags waste a lot of space. There is also a bug WRT iterators. Some kids in
the ref chain are actually on the stack only the tree is safe to edit because
the stack variables are not full kids.

The duplicate ignores causes a complicated implementation in the print
function. It may actually be unneeded now that we have follow ignores. We can
also provide some control over where the ignore tokens go, to the right of the
left token or to the left of the right token.