added the trim operation (^)
The trim operation wraps the ignores with nodes that have the suppress-left and
suppress-right flags set in the ignore list struct. These are now added to the
list of ignores to ouput once a terminal is hit, but are not printed since they
have no data. They are there just for their flags. Implemented the
suppress-left and suppress-right operations in a walk of the ignores to output
by altering the list.