added trim control flag to print code, auto-trimming all colm print calls
The print implemenation now takes a trim flag. The colm print function now sets
this flag by default. This is a change to the colm language back to 0.5
semantics. The $ conversion uses this flag too (also 0.5 semantics), in the
previous commit it issued a tree trim operation. The % operation gives a string
conversion without triming.