improvements to ignore handling in the parser
Every region now also has a duplicate scanning region that is only for tokens.
The duplicate ignores and tokens generate the original tokens through a
TokenDef ignore mechanism.

Can turn off post ignore parsing and pre-igore parsing on a token-by-token
basis. Probably want to move it into the productions and specify it there.
Currently don't have a specification mechanism.

If an ignore is a post-token ignore it is not right-attached.