experimenting with use of a nonterm for collecting ignores.
Can say that a production should collect ignores from a region. There is a
collect ignore region created, but the states from the ignore-version of the
region is used. When the scanner fails to produce a token from the
collect-ignore region, the collect-ignore token is generated and accepted by
the fsm. Need to take it out of the data tree on reductions and put it into an
ignore list. Reverse this during unparsing.