a new parse statement syntax
Added a parse statement constructs the parser, sends some text in the style of
the send statement, then returns the parser. More can be sent. When done finish
is called. The goal here is to eliminate the parse statements that return the
parsed tree because there is no way to get the parse error. The parser is
immediately destroyed.