stack bounds checking, not in effect yet
Added bounds checking on the stack pointer. Not in effect yet since the bounds
are not set. Will grow and shrink as necessary. Need to eliminate the
vm_top_off macro. It is not possible to use it with an incontiguous stack
unless all the pushes and pops it span a single push/pop. Where it is used
(parsing) that won't be the case.