keep a reserve stack block around for balanced push/pop following contiguous call
There may be some balanced push/pop calls after a call to contiguous, but we
still need the contiguous space there. The pop can cause the contiguous stack
block to be freed. One way to prevent this is to keep the last popped block
around as a reserve and use it if it's big enough. Not sure if this is good
enough, but so far seems to work ok.