eliminated ranges from condition transitions
Condition values benefit very little from range-base orientation. We likely
have groups of values with bits both on and off, but only the lowest bit makes
the values adjacent in a numerical ordering. Some other ordering that lends
itself to optimization may be found, but leaving that to some later date (once
the current implementation is functional).