started on the table-array abstraction
The TableArray class will be used to collect information on tables, then to
generate them. We will feed the data in two passes, both times with the same
interface. What is done with the data (analyse or generate) will be determined
by the class's state. The analysis pass is there to discover the limits, choose
a type for the array and allow us to predict its size.