disabled all code generators except C
Will now eliminate the data structure support for the original condition
implementation. Will also commit to the new implementation in the sense that we
will eliminate the fallback to the non-condition transitions. The target state
and actions will be removed from the reduced transition data structure. This
will force the conversion of the code generators to the inner condition style.