removed pointers from FsmRun back to attached streams
These pointers are not necessary. "Attached" means "FsmRun holds the data of" a
stream. Suffices to point from stream to FsmRun and when data is copied into an
FsmRun to check if there is an other FsmRun holding the stream's data. At that
point the detach can occur.

Also unified the two stream attach/detach functions.