parse expression returns the tree type, added 'error' global
Reverted back to the original semantics of the parse expression, where the
parser object is a temporary and the tree is returned. It is inconsistent form
to use:

parse Label: type [input]

and have Label be of something other than 'type'.

Added a global variable called 'error', which stores the error of the most
recently executed parse statement. This lets us have the better semantics for
the concise parse expression, and be able to get at the error.