use the #ifdef syntax for OUTPUT in lang-indep tests
This now matches the runtest syntax for output. The ifdef syntax was
inconvenient with the TXL lang trans because of a conflict with ragel comments,
so C comments were used. Now that we want to parse C comments to extract
variables, it is better to use the ifdef syntax and we can do it more easily in
colm.