barracuda patch: mmap load indicies
We were having issues with long compile times, and determined that the large
static arrays generated by ragel significantly slowed down gcc. To solve this
issue, we created the mmap_load_indices.patch attached. The patch adds a flag
to ragel, --indexfileprefix, that takes an optional argument such as
--indexfileprefix=3D"/dev/shm/prefix_". This flag will cause ragel to generate
a binary file of the indices array instead of hard coding them into the .c
file. The patch then adds two write statements that ragel recognizes, %%write
include, and %%write load_index, to facilitate the loading of this file at
run time. The write include statement adds the proper #include statements so
that the necessary libraries are loaded. The write load_index statement makes
a function call that actually loads the file. (The function is spit out with
%%write data). With this patch we are able to compile files that previously
took an indefinite amount of time.