barracuda: progress bar
Adds a progress bar to ragel, using the -r flag. When ragel is run with -r, a
progress bar is printed that shows how far we've walkedin the parse tree. The
algorithm I use to do this is very simple. I first count the number of nodes in
the tree. Then, we as we walk the tree, percentage as current_node_count /
(number_nodes*2) is output, where current_node_count is incremented twice for
each node. Once on entrance and once on exit. I also add a trail of
"breadcrumbs" to the output so that you can see what rule you're currently if
ragel is taking forever.