added call to rlhc as final step to compilation
Going to experiment with a final compilation step that transforms generated
code into the target host language. This differs from the rlcodegen of ragel 5
in that the final step is receiving code, not a declarative FSM.

The intermediate code will be an invented language that is only rich enough to
express the programming language constructs needed to implement a state
machine.