added backtick lexical syntax for literal tokens
Added a new syntax for literal tokens. A single backtick is starts the literal.
Ends at whitespace or ] in second or greater position.

The zero length token is now just a plain backtick.

Removed comma as separator in literal statement.

Single-quoted strings are now treated like double-quoted strings. They may
appear in expressions and are concatenated with the other string forms.

Allowing ',' in parameter and argument lists to separate arguments,
eg foo('a', 'b').