removed comma operator from expression lists
If comma is allowed there will likely be large amounts of code written with it
because it's so comon in other languages. It is only needed in a few cases, so
for those we can resort to other techniques, such as using parens around
expressions to isolate ('a') ('b'). This will prevent confusion when seeing
code with and without commas.