forked off fncall, fnret and fnbreak
These will implement the call/ret/break functionality in the style of fnext.
They will not jump to the target state, but rather cause the change on the next
character iteration. They can be implmented without use of gotos or exceptions
in the target language. These will be alternatives for languages that don't
support jumping.