renaming to reflect integration of non-goto (var) drivers
Languages that don't have goto, nor any construct allowing them to be faked for
the purpose of re-entering the main loop will use non-goto versions of most of
the code generators. They use variables to control which blocks to run. Did
some renaming to allow the introduction of these drivers.