moved default transitions out of indicies array
This was cherry-picked from the flat-interleave branch. Removing the default
transition from the indicies will simplify the the character equiv class
driver. We want to keep low-key and high-key in the orig keyspace, so the key
span will no longer be a way of finding the default transition at the end of
the indicies. refs #22