removed slen from the flat driver
Removed the index into the trans length array from the flat driver. Now that
defaults are no longer in the main index array we were using the value only to
determine if the transition array was empty. We can use keys for that by
emitting an impossible range for that case. Overall a win. refs #22.

mailbox1    78.167 -> 77.429  1.00953
strings1    72.067 -> 71.337  1.01023
strings2    76.675 -> 68.812  1.11427
rlscan      64.724 -> 56.743  1.14065
cppscan1    76.066 -> 74.173  1.02552