the pair iterator must now be passed an iterator type
Now requiring that the type arg to pair iterator be one of the custom iterators
written for PairIter. Implemented PiList, which wraps the list pointer. This
will allow us to swap in an iterators for vector so we can cross a list and a
vector, or a vector and a vector. refs #22