split the transition struct into base and implementation
The base class holds the key data and the condition class. The imple class
holds the list of condition transitions. This will allow us to substitute in an
implementation class that holds transtion data itself. When a transition has no
condition space we we will not be forced to pay the memory cost for condition
lists. refs #31