trans.lm: added a concrete grammar for expressions
Was using an abstract grammar that did not capture precedence. This was fine
for all existing host language transformations, but not for ASM. No parsing in
the concrete and transforming to abstract before writing out. Frees us from
having to walk the concrete grammar (more work) in every language translator.