added state push/pop actions to nfa-repeat construction
Can now specify push/pop actions to run when pushing and popping NFA choice
points. These can be used to save and restore the state necessary for counting.
The actions become attached to the nfaOut entries, then to the nfaTargs in the
reduced machine and are generated like to or from actions are generated. The
push is executed when the choice point is stored. The pop is stored in the same
record and executed when the choice point is popped.

refs #47