removed the name from the nfa union and nfa union terms
The top level name was only for debug output. This can be replaced with a print
of the machine name. The term name was for selecting the term to check in the
full output. This is no longer necessary since we generated a union with a
single term when when we check now.

refs #43