first cut at a cond-space solution to out conditions
Doing something similar to what is done in transition structure.

A couple problems:

1. The resulting data structure is not well suited to embedding leaving
conditions since it is key based, which is designed for evaluation, not
embedding.

2. This is a union operation. We need to distinguish between union and
intersection. In merges that are a result of concatenation, we need to use
intersection.

Probably need to keep condition state. Then a vector for each position that
tracks the senses. This would be better suited to the embedding, also can use
it more easily for intersection.

refs #52