implemented a case-independent range operator
Had been attempting this in p2r, but it's hard to do there considering the many
forms of range ends. In ragel, we have the ends converted to integers and it
can be done once, for all kinds of end specification forms. Fixes a test case
failure in the PCRE suite.