split the target when following epsilon transitions
When embedding actions and conditions and we follow epsilons, we need to split
the target for a 100% solution. This solves some immediate failures. But it
needs a more general solution. Alternative is to avoid the problem by
transferring actions and conditions to epsilon transitions. That is the cleaner
solution with much less code touched. refs #64