first cut of transfer of out conds/actions to nfa transition
This approach replaces the following of epsilong transitions when embedding
actions and conditions. It is a much cleaner, localized solution, that avoids
some difficult problems with a fixed point iteration deep inside the core ragel
operations.

refs #47

This first cut creates a significant perfomance hit that needs to be fixed with
a better runtime implementation.