successful two-pass parser that will allow online parsing
Where previously we would parss all sections before processing, we now make a
first pass to collect the list of sections and track last references. We then
rewind the input file and reparse. We are now in the position to compile when
we encounter a last section on the second text pass.

refs #64