grammar re-org so we don't have to access children of reduced items (in .lm)
This change allows us to clean up data trees during reduction the same way we
clean up parse trees. Would be better if we didn't need to make that change.
Can likely be fixed by going one level deeper in the cleanup. Wouldn't be an
unreasonable compromise, so we can still pass data up in the colm reduction
actions.