declare global variables as constant in Julia
In Julia, static global variables should be declared with the `const` keyword if
possible. This is an important hint to the Julia compiler because it will make
type inference work well and hence improve the performance.

This patch was picked from the BioJulia/ragel7 mirror, originally written by
Daniel C. Jones.