don't use Token::append for communicating - or + in number text
Instead of using Token::append for communicating - or + in number text, pass a
bool up to the place where the number is converted to an integral type. This
eliminates one major use of append. After eliminating the other (optim of
concat) we can strike modification from the token type. This will simplify the
management of memory used for token data, since there are three sources (load,
kelbt reduce, colm-reduce).