manage file name allocations and allow them to be exported
File names used in locations need to live beyond a program delete, but
allocating them when exporting locations is too costly. Instead manage them in
the colm program and allow them to be exported so they can live beyond the colm
program. Will do something similar with the token text.