do the shifting of start action orders in the functions that require it
Moved the shifting of start action orders from parse tree to fsmgraph. They
belong there, but we previously kept the cur action ordering in the parse tree.
Now have it in the right place and can do the shifting in the right place as
well. Goal here is to make FsmAp more useful in the absence of ParseData.