ragel codegen: optionally do not generate entry point variables
Many projects (nginx and lua appear in a quick google search) have had
to jump through hoops in order to avoid "unused variable" warnings from
compilers such as clang.

Reported in https://www.mail-archive.com/ragel-users@complang.org/msg00977.html
with the response: "it should be an option".

This patch makes it an option. I added it to all of the codegens which seemed
to be outputting names that matched "en_", and followed the same pattern as
the noError and noFinal flags. Tested only in straight C. Applies to ragel 6.9.