auto-trim in print, send and construct, disable with @
Now automatically trimming when building trees and printing. This is consistent
with the $ operator. Rationale is to standardize on the most commonly desired
behaviour and provide options for the less-needed.

Also much less likely to result in surprising output. There are many cases
where test inputs won't have any surrounding ignores, making it easy to forget
a trim is required. Then the occasional tree with surrounding ignores will give
an undesired result. Better to standardize on trim, making the default
behaviour more consistent.

The trim can be turned off with the @ operator in front of an expression.