added a clear statement, testing as alternative to two-pass parsing
The two-pass parser adds quite a bit of complexity to ragel internals.
Experimenting with an explicit clear statement as an alternative to an initial
pass just so we can know when we are done with

Clear should be necessary only in situations where ragel files are very large
and containing multiple machines.