added an experimental condition bit eliminator
With increasing condition embedding it becomes possible that some conditions
don't need to be tested because both cases lead to the same transition being
taken. Find those and eliminate.

Not sure we would want this on by default. It may be desirable that we can rely
on conditions being executed. For example what if the user relies on some side
effect? Maybe it should be an option. Or we simply emphasize that conditions
may be optimized out.