perform include by pushing the file stream to the input
Performing include in the colm-based grammar by scanning for the include token,
then in a translate (lexical code) block bringing in the rest of the include
statement with a separate parse statement, opening the file, and pushing the
file to the input stream.

Also now doing lookup and storage of actions with parameters in the colm
parsing program.