use BUILT_SOURCES for colm-generated code
Ragel 7 method is to use BUILT_SOURCES, but not include generated code in the
dist file. In ragel 6 we included the generated parsers so that building from
the tarball didn't require an existing version of ragel installed, but now we
are not depending on ragel to build ragel and need to have a built-installed
version of colm anyways for linking, so also generate the code using the colm
program.