moved eof transition execution to after conds+actions
The eof transition, as currently implemented, is only used as part of scanners.
Move the execution of this transition to after the conditions and actions.

Rationale for this is that eof actions are part of the regex comprising the
pattern and are therefore "inner".

Completing this change likely requires some changes to ensure that eof actions
are executed when we backup to abort an attempt at a longer pattern.