make it possible to set parser input before parsing
Make it possible to set the parser input (sourced from some other parser) the
same way the parser context is set. Previously would "send" the input the same
way a stream was sent, but this was inconsistent with how the system operates.
The "send" of an input was actually a hack that replaced the input. We are not
supporting recursing input structures. Just an input and underneath that
possibly some streams or data.