select between parser/plain stream at compile time, switch to "_"
Now that we have multiple parser using the same stream it is not responsible to
set a pointer to parser from stream and rely on that to decide where to send
(or to check if it is a simple stream). Back to making the decision at compile
time.

Also switching from "stds" to "_".