some array reference fixes (wrong array referenced, duplicates removed)
Sometimes used the wrong array for the value call, which didn't affect the
value, just type. Exposed once relying on the referenced bool. Added an assert
for ensuring we don't add values on arrays that are not being written to (not
started).

Also duplicted some arrays in subclasses, resulting in wrong notion of
referenced bit.