preserve original :nfa operator and use a new version for greedy and lazy version
The new version adds an extra step at the front where an immediate exit is
possible and the repeat is checked before entering the first time. This
provides better support for matching the zero-length word. It's possible to
init, then immediately exit if after init, it is determined that no matching
should take place (rep = 0).

Probably this new version should replace the previous one, though it breaks
usage of :nfa where we don't want any repetition beyond number one, and a
constant failure of rep is used for that. With this new version we don't get
any machines.

Those cases probably need to be migrated to a new NFA operator that does not
repeat at all, instead just wraps with NFA transitions so we can move through a
machine (no repetition) with NFA state protections.

Doing this would give us three kinds of NFA operations: union, repetition and a
wraper that can be used to concatenate two regions that require an NFA barrier.
Get this by wrapping a zero-length match. But can also use it to protect
capture. In any case, seems it will cover all our needs.