maintain reference counts when adding init ts actions
If there is a scanner in the graph we add some extra init ts actions to
to-state sets after the build. We do this to ensure ts is set to non-null only
when a token match is active.

This embedding happens after analysis because we need to know if there are
calls in actions. When we do this, keep the refcounts up to date so that
function call can stand on it's own and not mess up the refcounting. Not doing
it results in ugly crashes if tokstart is not already embedded.