always check P against EOF after regular and EOF actions
This change allows us to check if P has changed during action execution in both
regular and EOF cases. At this point we can decide if we want to exit. If we
don't hold in an EOF action we don't crash. If we alter P to = EOF in an action
we are also ok. Will allow simplification of the table based loop. Also is more
consistent, treating EOF more like a regular action.