merging aapl src dirs: added CmpStr and CmpString to colm/aapl
CmpStr is defined in colm/*.h so we need to also move it aside when we add it to aapl.

With this commit there are no more diffs between the aapl src dirs, with the
exception of a three files added to the aapl-repos package. Will use this as
the definitive version.