implemented compound expression and assignement expression
Placed these in the expression hierarchy and removed the assignment from the
top level. The docs seem to indicate that the result of an asignment can be
used in another assignment (chaining with ops evaluating right-to-left),
however the type system appears to prevent it as an assignment or compound
assignment returns ().