improvements to statement and expression to better match grammar
Previous version was made to work based on example code. This version is based
off the grammar in the rust-lang book and handles expressions with blocks,
which do not require semis in a statement list (did not handle this before),
and also the expression without block at the end of a statement list, without
requiring a semi (handled this before).