added test case demonstrating colm scanner failure
The comment pattern, which begins with '//', is interfering with the proper
scanning of '/' (div). When the longer token fails to match the scanner is not
executing the action for the shorter action.

Proper solution here is to move forward with replacing the colm scanner, which
is a clone of an early version of ragel, with the ragel library. Might make
sense to first fix the clone, provided it is an easy fix.