need to reset tokend when a pre-region scan fails
Following the scan of a token we attempt to scan just for ignore, before moving
on to the next region. This lets us collect ignores from the same region after
a token, but before moving on to the next region. But if the ignore parsing
fails we need to reset some vars before we attempt with the next region. Missig
here is the reset of toklen.