renamed toklen to tokpref and use it just for prefix
Had an interdependency between toklen and tokend. Was setting one from the
other and vice versa. Cleaned this up by renaming toklen to tokpref and using
it just to store the number of characters already matched before the current
buffer block. Tokend stays as tokend (in the ragel sense) and is used by the
parse to extract the match.