allow backtick literals to have an end quote.
Rather than exclude ? + and *, exclude ` from literals and allow backtick
literals to be end quoted.

Allows us to follow the literal with an operator without using a space between
the two, which is very unnatural. Can write: [`,`?].

If we forget to the end quote it we end up with literal that (probably) does
not exist. It should generate an error, rather than silently give us something
we don't want.

Note that ] is still excluded from literals in positions 2+.