loadcolm cannot be shared by bootstrap0 and colm
Attempting to do so results in the following circular dependency:

loadcolm.cc -> if3.h -> boostrap2 -> loadcolm.cc

Fix this by using the source file loadfinal.cc, which is not is not compiled
directly, but rather included in sources. It is used for both bootstrap2 and
colm, but these programs each use separate gen/if[23].h headers, so this file
is included by distinct source files that each bring in the appropriate
dependency, then include this file.