colm rewriting: newly generated prods should go above where they are created
Redid the sublist implementation in colm such that the new produciton is placed
immediately above the place where it it was first referenced. Works better in
the face of of namesapce usage

Also removed the repeat implementations. These are already correct in the C++
program so no sense in redoing them. This is in line with the notion that we
can use colm transformations for new language features, before they are
eventually moved into the main program.