Go codegen: test p VS pe in goto/call/ret in EOF actions
refs #78. This is the Go version of 284f1fb0ba7c2a5de6439761f0df0cc74f96e8b8:

If an fgoto, fcall or fret is issued in an EOF action check p VS pe before
continuing. If at the end of the block then re-issue the eof test.

If a program properly issues an fhold before the control flow the program won't
be affected. If the user neglects to do this we won't run off the end of the
buffer.