colm: added outdent check on literal concatenations
The colm language has an ambiguity between literal concatenations and distinct
statements, where the first ends in string/accum/cons/pat and the second is a
bare send. For example:

match P
	"a b c d"
"e f g h"

Warn on some of these by checking for an outdent (relative to the first
component) on top level concatenations. If the user does not outdent between
separate intended statements we cannot help with this check.

refs #95